refactor: ETF 보유 주수 타입을 Int에서 Double로 변경

+++ frontend/WYE/app/src/main/java/com/d102/wye/domain/model/PortfolioDetail.kt
@@ -13,5 +13,5 @@ data class PortfolioDetail(
 
 data class PortfolioCount(
     val ticker: String,
-    val counts: Int
+    val counts: Double
 )
